Implemented different routes for the ZoosDB and BearsDB.
